added scripts to switch content
