+++ course-101/android/app/src/main/java/live/ditto/quickstart/tasks/data/DittoManager.kt
@@ -40,6 +40,24 @@ class DittoManager(
             //TODO: setup Ditto Identity
             //
             //UPDATE CODE HERE
+            val identity = DittoIdentity.OnlinePlayground(
+                androidDittoDependencies,
+                dittoConfig.appId,
+                dittoConfig.authToken,
+                false,
+                dittoConfig.authUrl
+            )
+            this.ditto = Ditto(androidDittoDependencies, identity)
+
+            //
+            //TODO:  setup transports for P2P
+            //
+            // Set the Ditto Websocket URL
+            ditto?.updateTransportConfig { config ->
+                config.connect.websocketUrls.add(dittoConfig.websocketUrl)
+                // TODO: enable all P2P transports
+                config.enableAllPeerToPeer()
+            }
 
             // disable sync with v3 peers, required for syncing with the Ditto Cloud (Big Peer)
             this.ditto?.disableSyncWithV3()
@@ -89,6 +107,17 @@ class DittoManager(
                     //
 
                     //UPDATE CODE HERE
+                    ditto?.store?.execute(
+                        "INSERT INTO tasks INITIAL DOCUMENTS (:task)",
+                        mapOf(
+                            "task" to mapOf(
+                                "_id" to task._id,
+                                "title" to task.title,
+                                "done" to task.done,
+                                "deleted" to task.deleted,
+                            )
+                        )
+                    )
                 } catch (e: Exception) {
                     errorService.showError("Unable to insert initial documents: ${e.message}")
                     Log.e(TAG, "Unable to insert initial document", e)
@@ -121,14 +150,19 @@ class DittoManager(
             //
 
             //UPDATE CODE HERE
-            val query = ""
+            val query = "SELECT * FROM tasks WHERE NOT deleted"
 
             //
             //TODO: - setup store observer with query and set array with TaskModel
             //
 
             //UPDATE CODE HERE
-
+            storeObserver = ditto?.store?.registerObserver(query) { results ->
+                val items = results.items.map { item ->
+                    TaskModel.fromMap(item.value)
+                }
+                trySend(items)
+            }
             awaitClose {
                 //
                 //TODO: - Clean up the observer when the flow is cancelled
@@ -207,6 +241,7 @@ class DittoManager(
                     //
 
                     //UPDATE CODE HERE
+                    it.startSync()
 
                     //
                     //TODO: implement the set subscription
@@ -214,6 +249,8 @@ class DittoManager(
                     //
 
                     //UPDATE CODE HERE
+                    val subscriptionQuery = "SELECT * from tasks"
+                    subscription = it.sync.registerSubscription(subscriptionQuery)
                 }
             } catch (e: Exception) {
                 errorService.showError("Failed to start ditto sync: ${e.message}")
@@ -272,7 +309,7 @@ class DittoManager(
                 //
 
                 //UPDATE CODE HERE
-                val query = ""
+                val query = "INSERT INTO tasks DOCUMENTS (:newTask)"
 
                 //
                 //TODO: use dittoInstance store to execute DQL with arguments
@@ -280,7 +317,17 @@ class DittoManager(
                 //
 
                 //UPDATE CODE HERE
-
+                ditto?.store?.execute(
+                    query,
+                    mapOf(
+                        "newTask" to mapOf(
+                            "_id" to taskModel._id,
+                            "title" to taskModel.title,
+                            "done" to taskModel.done,
+                            "deleted" to taskModel.deleted
+                        )
+                    )
+                )
             } catch (e: Exception) {
                 errorService.showError("Failed to add taskModel: ${e.message}")
                 Log.e(TAG, "Failed to add taskModel:", e)
@@ -306,7 +353,13 @@ class DittoManager(
                 //
 
                 //UPDATE CODE HERE
-                val query = ""
+                val query = """
+                UPDATE tasks
+                SET title = :title,
+                    done = :done,
+                    deleted = :deleted
+                WHERE _id = :_id 
+                """
 
                 //
                 //TODO: use dittoInstance store to execute DQL with arguments
@@ -314,7 +367,15 @@ class DittoManager(
                 //
 
                 //UPDATE CODE HERE
-
+                ditto?.store?.execute(
+                    query,
+                    mapOf(
+                        "title" to taskModel.title,
+                        "done" to taskModel.done,
+                        "deleted" to taskModel.deleted,
+                        "_id" to taskModel._id
+                    )
+                )
             } catch (e: Exception) {
                 errorService.showError("Failed to update taskModel: ${e.message}")
                 Log.e(TAG, "Failed to update taskModel:", e)
@@ -356,7 +417,11 @@ class DittoManager(
                     //
 
                     //UPDATE CODE HERE
-                    val query = ""
+                    val query = """
+                        UPDATE tasks
+                        SET done = :done 
+                        WHERE _id == :_id
+                    """
 
                     //
                     //TODO: use dittoInstance store to execute DQL with arguments
@@ -364,6 +429,13 @@ class DittoManager(
                     //
 
                     // UPDATE CODE HERE
+                    it.store.execute(
+                        query,
+                        mapOf(
+                            "done" to !done,
+                            "_id" to id
+                        )
+                    )
                 }
             } catch (e: Exception) {
                 errorService.showError("Failed to update taskModel: ${e.message}")
@@ -389,9 +461,17 @@ class DittoManager(
                 // https://docs.ditto.live/sdk/latest/crud/delete#soft-delete-pattern
                 //
                 //UPDATE CODE HERE
-                val query = ""
-
-
+                val query = """
+                UPDATE tasks
+                SET deleted = true
+                WHERE _id = :_id 
+                """
+                ditto?.store?.execute(
+                    query,
+                    mapOf(
+                        "_id" to id,
+                    )
+                )
             } catch (e: Exception) {
                 errorService.showError("Failed to archive taskModel: ${e.message}")
                 Log.e(TAG, "Failed to archive taskModel:", e)
